Adopt Chrome's interface changes. This is the the NaCl version of http://codereview.chromium.org/6711047/
BUG=none
TEST=NaCl compiles with new Chrome version.

Review URL: http://codereview.chromium.org/6722004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4669 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 0d8d5ba7eb9ccf5b12e4f53396469532f91fa512

+++ src/shared/ppapi_proxy/plugin_ppb_core.cc
@@ -1,4 +1,4 @@
-// Copyright 2010 The Native Client Authors. All rights reserved.
+// Copyright 2011 The Native Client Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can
 // be found in the LICENSE file.
 
@@ -59,7 +59,7 @@ void ReleaseResource(PP_Resource resource) {
   }
 }
 
-void* MemAlloc(size_t num_bytes) {
+void* MemAlloc(uint32_t num_bytes) {
   DebugPrintf("PPB_Core::MemAlloc: num_bytes=%"NACL_PRIuS"\n",
               num_bytes);
   return malloc(num_bytes);
